backend/requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP client and utilities
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.0

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
pypdf==3.17.1
pdfplumber==0.10.3
pymupdf==1.23.8
pillow==10.1.0
opencv-python==4.8.1.78

# LangChain ecosystem
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.0
langchain-experimental==0.0.47
langgraph==0.0.20
langserve==0.0.30
langsmith==0.0.69

# Multimodal and document processing
unstructured[all-docs]==0.11.6
layoutparser[paddleocr]==0.3.4
paddlepaddle==2.5.2
paddleocr==2.7.0.3
easyocr==1.7.0
pytesseract==0.3.10
pdf2image==1.16.3
tabula-py==2.8.2
camelot-py[cv]==0.11.0

# Computer vision and image processing
scikit-image==0.22.0
matplotlib==3.8.2
seaborn==0.13.0

# Document layout analysis
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
transformers==4.36.2
torch==2.1.1
torchvision==0.16.1

# Advanced PDF parsing and chunking
pdfminer.six==20231228
smart-pdf==0.1.0
deeplake==3.8.15
sentence-transformers==2.2.2

# Vector database and embeddings
chromadb==0.4.18
faiss-cpu==1.7.4
pinecone-client==2.2.4
weaviate-client==3.25.3

# NLP and NER tools
spacy==3.7.2
spacy-transformers==1.3.4
nltk==3.8.1
textstat==0.7.3

# Text processing and chunking
tiktoken==0.5.2
langchain-text-splitters==0.0.1

# Legal APIs and external integrations
westlaw-api==1.0.0  # Placeholder - actual package may vary
lexisnexis-api==1.0.0  # Placeholder - actual package may vary

# Additional NLP libraries
regex==2023.10.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Background tasks and caching
celery==5.3.4
redis==5.0.1

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0

# Environment and configuration
python-decouple==3.8